move out operator tables builds, add interval_arithmetic_operator_table
