+++ src/operator_descr.rs
@@ -3,106 +3,6 @@ use std::collections::HashMap;
 use crate::expression::Operation;
 use crate::semantics::*;
 
-pub fn default_operator_table() -> OperatorTable<f64> {
-    let unary_ops: Vec<UnaryOp<f64>> = vec![
-        UnaryOp {
-            symbol: "ln",
-            semantics: |x| x.ln(),
-        },
-        UnaryOp {
-            symbol: "log",
-            semantics: |x| x.log(10.0),
-        },
-        UnaryOp {
-            symbol: "sin",
-            semantics: |x| x.sin(),
-        },
-        UnaryOp {
-            symbol: "cos",
-            semantics: |x| x.cos(),
-        },
-        UnaryOp {
-            symbol: "tan",
-            semantics: |x| x.tan(),
-        },
-        UnaryOp {
-            symbol: "sqrt",
-            semantics: |x| x.sqrt(),
-        },
-        UnaryOp {
-            symbol: "-",
-            semantics: |x| (-x),
-        },
-        UnaryOp {
-            symbol: "sgn",
-            semantics: |x| x.signum(),
-        },
-        UnaryOp {
-            symbol: "abs",
-            semantics: |x| x.abs(),
-        },
-        UnaryOp {
-            symbol: "asin",
-            semantics: |x| x.asin(),
-        },
-        UnaryOp {
-            symbol: "acos",
-            semantics: |x| x.acos(),
-        },
-        UnaryOp {
-            symbol: "atan",
-            semantics: |x| x.atan(),
-        },
-    ];
-
-    let binary_ops: Vec<BinaryOp<f64>> = vec![
-        BinaryOp {
-            symbol: "+",
-            semantics: |x, y| x+y,
-            assoc: Assoc::Left,
-            prec: 1,
-        },
-        BinaryOp {
-            symbol: "-",
-            semantics: |x, y| x-y,
-            assoc: Assoc::Left,
-            prec: 1,
-
-        },
-        BinaryOp {
-            symbol: "*",
-            semantics: |x, y| x*y,
-            assoc: Assoc::Left,
-            prec: 2,
-        },
-        BinaryOp {
-            symbol: "/",
-            semantics: |x, y| x/y,
-            assoc: Assoc::Left,
-            prec: 2,
-        },
-        BinaryOp {
-            symbol: "^",
-            semantics: |x, y| x.powf(y),
-            assoc: Assoc::Left,
-            prec: 3,
-        },
-    ];
-
-    let consts: Vec<ConstantOp<f64>> = vec![
-        ConstantOp {
-            symbol: "pi",
-            semantics: std::f64::consts::PI,
-        },
-        ConstantOp {
-            symbol: "e",
-            semantics: std::f64::consts::E,
-        },
-    ];
-
-    OperatorTable::new(unary_ops, binary_ops, consts)
-}
-
 // Description of an operator supported
 // For now we support three types of operators:
 // - Unary operators, are written in prefix form
